% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{print.bluetooth_data}
\alias{print.bluetooth_data}
\title{Print Method for Bluetooth Data}
\usage{
\method{print}{bluetooth_data}(x, ...)
}
\arguments{
\item{x}{A bluetooth_data object.}

\item{...}{Additional arguments passed to print.data.frame.}
}
\description{
Print Method for Bluetooth Data
}
