% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_interface.R
\name{summarize_device_counts}
\alias{summarize_device_counts}
\title{Summarize Device Detection Counts}
\usage{
summarize_device_counts(data)
}
\arguments{
\item{data}{A data.frame from \code{\link{process_bluetooth}}.}
}
\value{
A data.frame with summary statistics by device.
}
\description{
Generate summary statistics for each device across all timestamps.
}
