% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsbluetoothtracker-package.R
\docType{package}
\name{wsbluetoothR-package}
\alias{wsbluetoothR}
\alias{wsbluetoothR-package}
\title{wsbluetoothR: Process WildlifeSystems bluetooth data}
\description{
Process Bluetooth data from ws-bluetooth
}
\author{
\strong{Maintainer}: Ed Baker \email{ed@ebaker.me.uk}

}
\keyword{internal}
